requirements.txt
Flask==2.1.3
matplotlib==3.5.2
numpy==1.23.1
Pillow==9.2.0
torch==1.12.0
torchvision==0.13.0
Werkzeug==2.2.0
